tests/test_queue.py
"""Здесь надо написать тесты с использованием unittest для модуля queue."""
import unittest
from src.queue import Queue


class QueueTestCase(unittest.TestCase):
    def setUp(self):
        self.queue = Queue()

    def test_enqueue(self):
        self.queue.enqueue('data1')
        self.queue.enqueue('data2')
        self.queue.enqueue('data3')
        self.assertEqual(str(self.queue), "Queue: data1 -> data2 -> data3")

    def test_dequeue(self):
        self.queue.enqueue('data1')
        self.queue.enqueue('data2')
        self.queue.enqueue('data3')
        self.assertEqual(self.queue.dequeue(), 'data1')
        self.assertEqual(self.queue.dequeue(), 'data2')
        self.assertEqual(self.queue.dequeue(), 'data3')
        self.assertIsNone(self.queue.dequeue())

    def test_empty_queue(self):
        self.assertEqual(str(self.queue), "Queue: ")
        self.assertIsNone(self.queue.dequeue())


if __name__ == '__main__':
    unittest.main()